Fiddle factory
This seems close - I am seeing the message inside Game, but only the last message where I tried to pass a list of all messages, and I cannot find a way to write syntax to call the messageFunction. Close but no cigar. So near yet so far. Why is it so difficult to nail this?