More UI tweaks
The middle row had an unwanted gap under it. I change the escape hatch look. I added a border around the game board. I changed the tile rack look a little. I added code to prevent the outline on a clicked square from persisting. I fixed a bug where play not moving clicks a square and was getting an error.