Show game moves
Some changes to how chat looks and chat input behaves. Now allowing any characters because I realized the cursor positioning with the message being entered was lost, so I made it a bit less controlled and more normal for the user. I could not quite get it how I want. The UI is fine, but restricting characters was lost.