Break out player section into own file
Also refactored the code a bit to have a common routine instead of one for prisoners and one for guards. Also stopped auto-selecting a tile when player is typing the letters rather than clicking them on the rack and then clicking the intended square. Also, I changed how a selected tile rack displays and allow players to select when it is not their turn.
I had and fixed a problem of Build fails in playerSection.js where it did failed on things like props.racktiles.map, saying 'Cannot read property 'map' of undefined. I had to do "props.racktile && " in front of it. There were a couple of spots where props.moves.length needed props.moves && in front.