Add "One got away" sound byte, refactor arrays
The main change was going from squares[][] and usedby[][] to squareArray[][]. It was interesting that the array deep copy stopped working at this point, and trying to write a routine to build a deep copy did not quite work (the 'row' value got set to 8 on all rows, even though the other fields seemed fine). I settled on JSON.parse(JSON.stringify(array)) to make the deep copy.