Tidy up the lobby
My laptop was pushing the start button to the next row. I used w3-row outside the main area and chat, with w3-col s9 and w3-col s3 to make the chat use 25% of the width. That worked. I am doing the build and export now to get the previous two commits as well, since there is an actual difference to the user for this commit.