Add access control allow origin, call server route evtest, show value
This is a baby step. Server evtest gets a value that is not a secret but is from a secret place (my computer env vars locally, heroku config settings in prod). If this works it establishes a pattern for more complex things like getting game data from a data store.