Determine db update points
Change prisoners game start to send "startgame" instead of "hello", and to send all the game data. Put a console log in the server.js where it is the right time to update the data store for the game. Perhaps I could use my azure Scrabble club site to act as database API?