Possible bug fix
Had a game where prisoners rack was empty on turn 2, and it seemed related to web socket disconnects. I noticed an anomaly with 'senderid' vs 'thisisme' and fixed that, which may or may not fix this empty rack bug. I added code to tell the user what to do if it happens. I also turned off data API interaction since I am not using it yet for functionality.