Support periodic sync checks
This was much more difficult than anticipated, and my confidence is a little low on whether I truly got this correct. I ended up sending a request for a syncstamp from the front end (on an interval), then checking the syncstamp that the server sends and sending requestgamedata if out of sync. A bit convoluted but I could not get it to work with fewer steps.