Add examiners chat, fix bug, UI tweaks
You can now jump to a move in examiner mode by clicking on the move.
The bug fix was related to the move that emptied the tile bag - it was not setting tiles to empty so you saw those last picked tiles in the unseen tiles section.
The UI tweak was mainly the examiner tile rack arrow. It is now a more subtle floating bright green triangle (visible yet not UI greedy). I also added a hover to the move list so the examiners can see what they are potentially going to click.