Remove apology for slowness, fix bugs, add validate words call in server and dev mode test button for it
Secret word was showing 2 extra asterisks. Show Info was saying every word was valid. Test button for validate words call is in word info page, and uses alert to show the response text. It handles "?" or "." as wildcard letters. Technically it treats the comma separated word list as a list of regex patterns and it puts \b either side to enforce whole word matches. The intent is for prison break to call with a comma separated list of all words formed by a play and get info back, but I have not changed pb yet to call it.