FYB create game function
You can now create a game from the lobby, and it will make sure the game id is not already used. The game component will display the basics (game id, nickname, number of players). I put w3-mobile on the game UI so I want to see how it looks on my phone.