FYB free for all round
Free for all round implemented. Still no points system or way to end the game or "pass" a move. No time limits. At this point players can play the game without seeing points or being able to pass. It's kind of shareable. Also lobby does not see when a game is created ... still wondering whether to do that.