Transmogrify
You can now choose the number of moves and generate the puzzle. It will show you the start word and target word, and the minimum number of moves (which you chose in the first place). There is no mechanism to enter a solution in the site yet. You can do it in your head or on paper or whatever.