Transmogrify allow input of solution
Hit enter or click GO to enter next word after typing it in. It will validate that it is a valid anagram or swap or drop or insert, is not the same word, and is a valid word. It will congratulate you when it is solved and prevent input of more words. You may click to start a new puzzle anytime. No backsies - once a word is entered and accepted, you cannot take it back. You cannot restart the puzzle.